requirements.txt
paho-mqtt
plotly-express
pandas
